Refactor salary settings and self-service profile to support overtime and night differential requests
- Removed tax brackets configuration from salary settings view.
- Updated salary show view to remove tax brackets display.
- Enhanced self-service profile to include buttons for submitting overtime and night differential requests.
- Added modals for submitting overtime and night differential requests with appropriate forms.
- Implemented new routes for handling claim requests in self-service.
- Created tests for payroll tax bracket calculations and approved claim amounts for overtime and night differential.

+++ database/migrations/2026_05_18_100000_create_plotting_tables.php
@@ -14,8 +14,8 @@ return new class extends Migration
         if (!Schema::hasTable('employee_plottings')) {
             Schema::create('employee_plottings', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('employee_id');
-                $table->unsignedInteger('supervisor_id')->nullable();
+                $table->unsignedBigInteger('employee_id');
+                $table->unsignedBigInteger('supervisor_id')->nullable();
                 $table->date('date');
                 $table->decimal('amount', 14, 2)->default(0.00);
                 $table->timestamps();
